+++ packages/node/src/indexer/api.service.ts
@@ -91,72 +91,70 @@ export class ApiService
       logger.info('Using provided chain types');
     }
 
-    await Promise.all(
-      network.endpoint.map(async (endpoint, i) => {
-        const connection = await ApiPromiseConnection.create(
-          endpoint,
-          this.fetchBlocksBatches,
-          {
-            chainTypes,
-          },
-        );
+    for await (const [i, endpoint] of network.endpoint.entries()) {
+      const connection = await ApiPromiseConnection.create(
+        endpoint,
+        this.fetchBlocksBatches,
+        {
+          chainTypes,
+        },
+      );
 
-        const api = connection.unsafeApi;
+      const api = connection.unsafeApi;
 
+      this.eventEmitter.emit(IndexerEvent.ApiConnected, {
+        value: 1,
+        apiIndex: i,
+        endpoint: endpoint,
+      });
+
+      api.on('connected', () => {
         this.eventEmitter.emit(IndexerEvent.ApiConnected, {
           value: 1,
           apiIndex: i,
           endpoint: endpoint,
         });
-
-        api.on('connected', () => {
-          this.eventEmitter.emit(IndexerEvent.ApiConnected, {
-            value: 1,
-            apiIndex: i,
-            endpoint: endpoint,
-          });
+      });
+      api.on('disconnected', () => {
+        this.eventEmitter.emit(IndexerEvent.ApiConnected, {
+          value: 0,
+          apiIndex: i,
+          endpoint: endpoint,
         });
-        api.on('disconnected', () => {
-          this.eventEmitter.emit(IndexerEvent.ApiConnected, {
-            value: 0,
-            apiIndex: i,
-            endpoint: endpoint,
-          });
-        });
+      });
 
-        if (!this.networkMeta) {
-          this.networkMeta = connection.networkMeta;
+      if (!this.networkMeta) {
+        this.networkMeta = connection.networkMeta;
 
-          if (
-            network.chainId &&
-            network.chainId !== this.networkMeta.genesisHash
-          ) {
-            const err = new Error(
-              `Network chainId doesn't match expected genesisHash. Your SubQuery project is expecting to index data from "${
-                network.chainId ?? network.genesisHash
-              }", however the endpoint that you are connecting to is different("${
-                this.networkMeta.genesisHash
-              }). Please check that the RPC endpoint is actually for your desired network or update the genesisHash.`,
-            );
-            logger.error(err, err.message);
-            throw err;
-          }
-        } else {
-          const genesisHash = api.genesisHash.toString();
-          if (this.networkMeta.genesisHash !== genesisHash) {
-            throw this.metadataMismatchError(
-              'Genesis Hash',
-              this.networkMeta.genesisHash,
-              genesisHash,
-            );
-          }
+        if (
+          network.chainId &&
+          network.chainId !== this.networkMeta.genesisHash
+        ) {
+          const err = new Error(
+            `Network chainId doesn't match expected genesisHash. Your SubQuery project is expecting to index data from "${
+              network.chainId ?? network.genesisHash
+            }", however the endpoint that you are connecting to is different("${
+              this.networkMeta.genesisHash
+            }). Please check that the RPC endpoint is actually for your desired network or update the genesisHash.`,
+          );
+          logger.error(err, err.message);
+          throw err;
         }
+      } else {
+        const genesisHash = api.genesisHash.toString();
+        if (this.networkMeta.genesisHash !== genesisHash) {
+          throw this.metadataMismatchError(
+            'Genesis Hash',
+            this.networkMeta.genesisHash,
+            genesisHash,
+          );
+        }
+      }
 
-        endpointToApiIndex[endpoint] = connection;
-      }),
-    );
+      endpointToApiIndex[endpoint] = connection;
+    }
 
-    this.connectionPoolService.addBatchToConnections(endpointToApiIndex);
+    await this.connectionPoolService.addBatchToConnections(endpointToApiIndex);
     return this;
   }