+++ packages/node/src/indexer/indexer.manager.spec.ts
@@ -14,6 +14,7 @@ import {
   NodeConfig,
   ConnectionPoolService,
   StoreCacheService,
+  ConnectionPoolStateManager,
   MmrQueryService,
 } from '@subql/node-core';
 import { Sequelize } from '@subql/x-sequelize';
@@ -216,7 +217,10 @@ describe('IndexerManager', () => {
   xit('should be able to start the manager (v0.0.1)', async () => {
     indexerManager = createIndexerManager(
       testSubqueryProject_1(),
-      new ConnectionPoolService<ApiPromiseConnection>(),
+      new ConnectionPoolService<ApiPromiseConnection>(
+        nodeConfig,
+        new ConnectionPoolStateManager(),
+      ),
       nodeConfig,
     );
     await expect(indexerManager.start()).resolves.toBe(undefined);
@@ -227,7 +231,10 @@ describe('IndexerManager', () => {
   xit('should be able to start the manager (v0.2.0)', async () => {
     indexerManager = createIndexerManager(
       testSubqueryProject_2(),
-      new ConnectionPoolService<ApiPromiseConnection>(),
+      new ConnectionPoolService<ApiPromiseConnection>(
+        nodeConfig,
+        new ConnectionPoolStateManager(),
+      ),
       nodeConfig,
     );
     await expect(indexerManager.start()).resolves.toBe(undefined);